Range and domain geometry for (1,1) operators
