+++ Wrappers/Python/test/test_Operator.py
@@ -657,244 +657,8 @@ class TestGradients(CCPiTestClass):
         self.assertTrue( LinearOperator.dot_test(Grad3 , decimal=4, verbose=True))
 
 
-class TestBlockOperator(CCPiTestClass):
-    def setUp(self):
-        numpy.random.seed(1)
 
-
-    def test_BlockOperator(self):
-        M, N  = 3, 4
-        ig = ImageGeometry(M, N)
-        arr = ig.allocate('random')
-
-        G = GradientOperator(ig)
-        Id = IdentityOperator(ig)
-
-        B = BlockOperator(G, Id)
-        # Nx1 case
-        u = ig.allocate('random')
-        z1 = B.direct(u)
-
-        res = B.range_geometry().allocate()
-        #res = z1.copy()
-        B.direct(u, out=res)
-
-        self.assertBlockDataContainerEqual(z1, res)
-
-        z1 = B.range_geometry().allocate(ImageGeometry.RANDOM)
-
-        res1 = B.adjoint(z1)
-        res2 = B.domain_geometry().allocate()
-        B.adjoint(z1, out=res2)
-
-        self.assertNumpyArrayEqual(res1.as_array(), res2.as_array())
-
-        BB = BlockOperator( Id, 2 * Id)
-        B = BlockOperator( BB, Id )
-        v = B.domain_geometry().allocate()
-        B.adjoint(res,out=v)
-        vv = B.adjoint(res)
-        el1 = B.get_item(0,0).adjoint(z1.get_item(0)) +\
-              B.get_item(1,0).adjoint(z1.get_item(1))
-
-        self.assertNumpyArrayEqual(v.as_array(),vv.as_array())
-        # test adjoint
-
-        BB = BlockOperator( Id, 2 * Id)
-        u = ig.allocate(1)
-        z1 = BB.direct(u)
-        res = BB.range_geometry().allocate(0)
-        BB.direct(u, out=res)
-
-        self.assertNumpyArrayEqual(z1.get_item(0).as_array(),
-                                   u.as_array())
-        self.assertNumpyArrayEqual(z1.get_item(1).as_array(),
-                                   2 * u.as_array())
-        self.assertNumpyArrayEqual(res.get_item(0).as_array(),
-                                   u.as_array())
-        self.assertNumpyArrayEqual(res.get_item(1).as_array(),
-                                   2 * u.as_array())
-
-        x1 = BB.adjoint(z1)
-
-        res1 = BB.domain_geometry().allocate()
-        BB.adjoint(z1, out=res1)
-        self.assertNumpyArrayEqual(x1.as_array(),
-                                   res1.as_array())
-
-        self.assertNumpyArrayEqual(x1.as_array(),
-                                   5 * u.as_array())
-        self.assertNumpyArrayEqual(res1.as_array(),
-                                   5 * u.as_array())
-        #################################################
-
-        BB = BlockOperator( Id, 2 * Id, 3 * Id,  Id, shape=(2,2))
-        B = BB
-        u = ig.allocate(1)
-        U = BlockDataContainer(u,u)
-        z1 = B.direct(U)
-
-        self.assertNumpyArrayEqual(z1.get_item(0).as_array(),
-                                   3 * u.as_array())
-        self.assertNumpyArrayEqual(z1.get_item(1).as_array(),
-                                   4 * u.as_array())
-        res = B.range_geometry().allocate()
-        B.direct(U, out=res)
-        self.assertNumpyArrayEqual(res.get_item(0).as_array(),
-                                   3 * u.as_array())
-        self.assertNumpyArrayEqual(res.get_item(1).as_array(),
-                                   4 * u.as_array())
-
-
-        x1 = B.adjoint(z1)
-        # this should be [15 u, 10 u]
-        el1 = B.get_item(0,0).adjoint(z1.get_item(0)) + B.get_item(1,0).adjoint(z1.get_item(1))
-        el2 = B.get_item(0,1).adjoint(z1.get_item(0)) + B.get_item(1,1).adjoint(z1.get_item(1))
-
-        shape = B.get_output_shape(z1.shape, adjoint=True)
-        out = B.domain_geometry().allocate()
-
-        for col in range(B.shape[1]):
-            for row in range(B.shape[0]):
-                if row == 0:
-                    el = B.get_item(row,col).adjoint(z1.get_item(row))
-                else:
-                    el += B.get_item(row,col).adjoint(z1.get_item(row))
-            out.get_item(col).fill(el)
-
-        self.assertNumpyArrayEqual(out.get_item(0).as_array(),
-                                   15 * u.as_array())
-        self.assertNumpyArrayEqual(out.get_item(1).as_array(),
-                                   10 * u.as_array())
-
-        res2 = B.domain_geometry().allocate()
-        #print (res2, res2.as_array())
-        B.adjoint(z1, out = res2)
-
-        #print ("adjoint",x1.as_array(),"\n",res2.as_array())
-        self.assertNumpyArrayEqual(
-            out.get_item(0).as_array(),
-            res2.get_item(0).as_array()
-            )
-        self.assertNumpyArrayEqual(
-            out.get_item(1).as_array(),
-            res2.get_item(1).as_array()
-            )
-
-        B1 = BlockOperator(G, Id)
-        U = ig.allocate(ImageGeometry.RANDOM)
-        #U = BlockDataContainer(u,u)
-        RES1 = B1.range_geometry().allocate()
-
-        Z1 = B1.direct(U)
-        B1.direct(U, out = RES1)
-
-        self.assertBlockDataContainerEqual(Z1,RES1)
-        
-    def test_block_operator_1_1(self):
-        M, N ,W = 3, 4, 5
-        ig = ImageGeometry(M, N, W)
-        operator0=IdentityOperator(ig)
-        operator1=-IdentityOperator(ig)
-        K = BlockOperator(operator0, operator1, shape = (1,2))
-        bg=BlockGeometry(ig, ig)
-        data=bg.allocate('random', seed=2)
-        ans = K.direct(data)
-        #self.assertNumpyArrayEqual( ans.shape, ig.allocate(0).as_array())
-        self.assertNumpyArrayEqual( ans.as_array(), ig.allocate(0).as_array())
-        self.assertFalse(isinstance(ans, BlockDataContainer))
-        
-        ans2 = K.adjoint(ans)
-        self.assertTrue(isinstance(ans2, BlockDataContainer))
-        self.assertNumpyArrayEqual(ans2.shape, (2,1))
-        
-        range_data=ans.geometry.allocate('random', seed=2)
-        ans3=K.adjoint(range_data)
-        self.assertNumpyArrayEqual(ans3.shape, (2,1))
-        self.assertNumpyArrayEqual(ans3.get_item(0).as_array(), range_data.as_array())
-        self.assertNumpyArrayEqual(ans3.get_item(1).as_array(), -range_data.as_array())
-        
-        
-
-    @unittest.skipIf(True, 'Skipping time tests')
-    def test_timedifference(self):
-        M, N ,W = 100, 512, 512
-        ig = ImageGeometry(M, N, W)
-        arr = ig.allocate('random')
-
-        G = GradientOperator(ig, backend='numpy')
-        Id = IdentityOperator(ig)
-
-        B = BlockOperator(G, Id)
-
-
-        # Nx1 case
-        u = ig.allocate('random')
-        steps = [timer()]
-        i = 0
-        n = 10.
-        t1 = t2 = 0
-        res = B.range_geometry().allocate()
-
-        while (i < n):
-            steps.append(timer())
-            z1 = B.direct(u)
-            steps.append(timer())
-            t = dt(steps)
-            #print ("B.direct(u) " ,t)
-            t1 += t/n
-
-            steps.append(timer())
-            B.direct(u, out = res)
-            steps.append(timer())
-            t = dt(steps)
-            #print ("B.direct(u, out=res) " ,t)
-            t2 += t/n
-            i += 1
-
-        self.assertGreater(t1,t2)
-
-        steps = [timer()]
-        i = 0
-        #n = 50.
-        t1 = t2 = 0
-        resd = B.domain_geometry().allocate()
-        z1 = B.direct(u)
-        #B.adjoint(z1, out=resd)
-
-        while (i < n):
-            steps.append(timer())
-            w1 = B.adjoint(z1)
-            steps.append(timer())
-            t = dt(steps)
-            #print ("B.adjoint(z1) " ,t)
-            t1 += t/n
-
-            steps.append(timer())
-            B.adjoint(z1, out=resd)
-            steps.append(timer())
-            t = dt(steps)
-            #print ("B.adjoint(z1, out=res) " ,t)
-            t2 += t/n
-            i += 1
-
-
-    def test_BlockOperatorLinearValidity(self):
-        M, N  = 3, 4
-        ig = ImageGeometry(M, N)
-        arr = ig.allocate('random', seed=1)
-
-        G = GradientOperator(ig)
-        Id = IdentityOperator(ig)
-
-        B = BlockOperator(G, Id)
-        # Nx1 case
-        u = ig.allocate('random', seed=2)
-        w = B.range_geometry().allocate(ImageGeometry.RANDOM, seed=3)
-        w1 = B.direct(u)
-        u1 = B.adjoint(w)
-        self.assertAlmostEqual((w * w1).sum() , (u1*u).sum(), places=5)
-
+ 
 
 class TestOperatorCompositionSum(unittest.TestCase):
     def setUp(self):
